Add playbooks to setup and teardown Consul
Consul will be started in a Podman container and will be managed by
a systemd service. The playbook creates a container using Podman
and generates a systemd unit file to manage the container.

The container is run as root since I could not figure out how to
start rootless container as a systemd service. Jobs under
systemd's user scope cannot depend on jobs in system scope. This
lead to the service starting before pre-reqs are met. I also ran
into problems with mapping user namespaces. This requires a deep
dive.