Add playbooks to install and teardown postgres
Postgres will be started in a Podman container and will be managed
by systemd service. The playbook creates a container using Podman
and generates a systemd unit file to manage the container.

As part of the install process, a new system account with name
"postgres" will be created along with a folder in the home dir to
hold the data. This directory will be mounted to the container's
/var/lib/postgres/data. The systemd unit file will be created in
user's scope.

The container is run as rootless container under user postgres.